requirements.txt
cython
pytorch
opt_einsum
tqdm
matplotlib